Update Array Cardio Day 1 exercise
The code for this JavaScript exercise called 'Array Cardio Day 1' gets modified. The changes improve the readability of the code by organizing the content on more lines and adjusting the structure of the code. Moreover, the changes include additional features, such as filtering the list based on certain conditions, sorting the inventors by birthdate and years lived, and providing more exercises including sorting the people alphabetically and reducing an array of data.