Update CSS and JavaScript in calculator project
The CSS and JavaScript files in the calculator project have been modified. CSS changes involve converting usage of px to em for better scaling and refactoring #container and button layouts using grid and flex. In JavaScript, initial values have been set for variables num1, num2, operator, and text. The text variable is also now reset to "0" instead of an empty string when the clear function is invoked. The changes make the code cleaner, more consistent, and therefore easier to maintain.